requirements.txt
openai==1.12.0
chromadb==0.4.22
httpx==0.26.0
beautifulsoup4==4.12.3
python-dotenv==1.0.1
tiktoken==0.5.2
pydantic==2.5.3
